+++ coloraide/color/convert.py
@@ -6,37 +6,78 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:  # pragma: no cover
     from ..color import Color
 
+# If a set of poorly created color spaces have been registered that create a loop or
+# cannot be converted to an XYZ base, conversion chains could grow to an infinite size.
+# Provide an arbitrary, worse case limit to prevent a deadlock and inevitable crash.
+MAX_CHAIN_SIZE = 10
+
 
 def convert(color: 'Color', space: str) -> Vector:
     """Convert the color coordinates to the specified space."""
 
     if color.space() != space:
-        convert_to = '_to_{}'.format(space.replace('-', '_'))
-        convert_from = '_from_{}'.format(color.space().replace('-', '_'))
-
         obj = color.CS_MAP.get(space)
         if obj is None:
             raise ValueError("'{}' is not a valid color space".format(space))
 
-        # See if there is a direct conversion route
-        func = None
-        # Don't send NaNs
+        # Create a worse case conversion chain from XYZ to the target
+        temp = obj
+        count = 0
+        from_color = []
+        from_color_index = {}
+        name = ''
+        while name != 'xyz':
+            from_color.append(temp)
+            name = temp.space()
+            from_color_index[name] = count
+            temp = color.CS_MAP[temp.BASE]
+
+            count += 1
+            if count == MAX_CHAIN_SIZE:  # pragma: no cover
+                raise RuntimeError(
+                    'Conversion chain reached max size of {} and has terminated to avoid an infinite loop'.format(
+                        count
+                    )
+                )
+
+        # Start converting coordinates until we either match a space in the conversion chain or bottom out at XYZ
         coords = util.no_nans(color.coords())
-        if hasattr(color._space, convert_to):
-            func = getattr(color._space, convert_to)
-            coords = func(color, coords)
-        elif hasattr(obj, convert_from):
-            func = getattr(obj, convert_from)
-            coords = func(color, coords)
+        current = type(color._space)
+        if current.space() != 'xyz':
+            count = 0
+            while current.space() not in from_color_index:
+                # Convert to color's base
+                base_space = color.CS_MAP[current.BASE]
+                coords = current.to_base(coords)
 
-        # See if there is an XYZ route
-        if func is None and color.space() != space:
-            func = getattr(color._space, '_to_xyz')
-            coords = func(color, coords)
+                # Convert to XYZ, make sure we chromatically adapt to the appropriate white point
+                if base_space.space() == 'xyz':
+                    coords = color.chromatic_adaptation(current.WHITE, base_space.WHITE, coords)
 
-            if space != 'xyz':
-                func = getattr(obj, '_from_xyz')
-                coords = func(color, coords)
+                # Get next color in the chain
+                current = base_space
+
+                count += 1
+                if count == MAX_CHAIN_SIZE:  # pragma: no cover
+                    raise RuntimeError(
+                        'Conversions reached max iteration of {} and has terminated to avoid an infinite loop'.format(
+                            count
+                        )
+                    )
+
+        # If we still do not match start converting from the point in the conversion chain
+        # where are current color resides
+        if current.space() != space:
+            start = from_color_index[current.space()] - 1
+
+            # Convert from XYZ, make sure we chromatically adapt from the appropriate white point
+            if current.space() == 'xyz':
+                coords = color.chromatic_adaptation(current.WHITE, from_color[start].WHITE, coords)
+
+            for index in range(start, -1, -1):
+                current = from_color[index]
+                coords = current.from_base(coords)
+
     else:
         # Nothing to convert, just pass values as is
         coords = color.coords()