+++ coloraide/spaces/din99o_lch.py
@@ -1,15 +1,10 @@
 """Din99o Lch class."""
 from ..spaces import RE_DEFAULT_MATCH
 from .lch.base import LchBase
-from .din99o import Din99o
 from .. import util
 import math
 import re
 from ..util import Vector, MutableVector
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:  # pragma: no cover
-    from ..color import Color
 
 ACHROMATIC_THRESHOLD = 0.0000000002
 
@@ -50,31 +45,20 @@ def lab_to_lch(lab: Vector) -> MutableVector:
 class Din99oLch(LchBase):
     """Lch D65 class."""
 
+    BASE = 'din99o'
     SPACE = "din99o-lch"
     SERIALIZE = ("--din99o-lch",)
     DEFAULT_MATCH = re.compile(RE_DEFAULT_MATCH.format(color_space='|'.join(SERIALIZE), channels=3))
     WHITE = "D65"
 
     @classmethod
-    def _to_din99o(cls, parent: 'Color', lch: Vector) -> MutableVector:
+    def to_base(cls, lch: Vector) -> MutableVector:
         """To Lab."""
 
         return lch_to_lab(lch)
 
     @classmethod
-    def _from_din99o(cls, parent: 'Color', lab: Vector) -> MutableVector:
+    def from_base(cls, lab: Vector) -> MutableVector:
         """To Lab."""
 
         return lab_to_lch(lab)
-
-    @classmethod
-    def _to_xyz(cls, parent: 'Color', lch: Vector) -> MutableVector:
-        """To XYZ."""
-
-        return Din99o._to_xyz(parent, cls._to_din99o(parent, lch))
-
-    @classmethod
-    def _from_xyz(cls, parent: 'Color', xyz: Vector) -> MutableVector:
-        """From XYZ."""
-
-        return cls._from_din99o(parent, Din99o._from_xyz(parent, xyz))