+++ coloraide/spaces/hsl/base.py
@@ -1,13 +1,9 @@
 """HSL class."""
 from ...spaces import Space, RE_DEFAULT_MATCH, FLG_ANGLE, FLG_OPT_PERCENT, GamutBound, Cylindrical
-from ..srgb.base import SRGB
 from ... import util
 import re
 from ...util import Vector, MutableVector
-from typing import Tuple, TYPE_CHECKING
-
-if TYPE_CHECKING:  # pragma: no cover
-    from ...color import Color
+from typing import Tuple
 
 
 def srgb_to_hsl(rgb: Vector) -> MutableVector:
@@ -59,6 +55,7 @@ def hsl_to_srgb(hsl: Vector) -> MutableVector:
 class HSL(Cylindrical, Space):
     """HSL class."""
 
+    BASE = "srgb"
     SPACE = "hsl"
     SERIALIZE = ("--hsl",)
     CHANNEL_NAMES = ("h", "s", "l", "alpha")
@@ -123,25 +120,13 @@ class HSL(Cylindrical, Space):
         return coords, alpha
 
     @classmethod
-    def _to_srgb(cls, parent: 'Color', hsl: Vector) -> MutableVector:
+    def to_base(cls, hsl: Vector) -> MutableVector:
         """To sRGB."""
 
         return hsl_to_srgb(hsl)
 
     @classmethod
-    def _from_srgb(cls, parent: 'Color', rgb: Vector) -> MutableVector:
+    def from_base(cls, rgb: Vector) -> MutableVector:
         """From sRGB."""
 
         return srgb_to_hsl(rgb)
-
-    @classmethod
-    def _to_xyz(cls, parent: 'Color', hsl: Vector) -> MutableVector:
-        """To XYZ."""
-
-        return SRGB._to_xyz(parent, cls._to_srgb(parent, hsl))
-
-    @classmethod
-    def _from_xyz(cls, parent: 'Color', xyz: Vector) -> MutableVector:
-        """From XYZ."""
-
-        return cls._from_srgb(parent, SRGB._from_xyz(parent, xyz))