+++ coloraide/spaces/hwb/base.py
@@ -1,14 +1,9 @@
 """HWB class."""
 from ...spaces import Space, RE_DEFAULT_MATCH, FLG_ANGLE, FLG_OPT_PERCENT, GamutBound, Cylindrical
-from ..srgb.base import SRGB
-from ..hsv import HSV
 from ... import util
 import re
 from ...util import Vector, MutableVector
-from typing import Tuple, TYPE_CHECKING
-
-if TYPE_CHECKING:  # pragma: no cover
-    from ...color import Color
+from typing import Tuple
 
 
 def hwb_to_hsv(hwb: Vector) -> MutableVector:
@@ -40,6 +35,7 @@ def hsv_to_hwb(hsv: Vector) -> MutableVector:
 class HWB(Cylindrical, Space):
     """HWB class."""
 
+    BASE = "hsv"
     SPACE = "hwb"
     SERIALIZE = ("--hwb",)
     CHANNEL_NAMES = ("h", "w", "b", "alpha")
@@ -103,49 +99,13 @@ class HWB(Cylindrical, Space):
         return coords, alpha
 
     @classmethod
-    def _to_xyz(cls, parent: 'Color', hwb: Vector) -> MutableVector:
-        """SRGB to XYZ."""
-
-        return SRGB._to_xyz(parent, cls._to_srgb(parent, hwb))
-
-    @classmethod
-    def _from_xyz(cls, parent: 'Color', xyz: Vector) -> MutableVector:
-        """XYZ to SRGB."""
-
-        return cls._from_srgb(parent, SRGB._from_xyz(parent, xyz))
-
-    @classmethod
-    def _to_srgb(cls, parent: 'Color', hwb: Vector) -> MutableVector:
-        """To sRGB."""
-
-        return HSV._to_srgb(parent, cls._to_hsv(parent, hwb))
-
-    @classmethod
-    def _from_srgb(cls, parent: 'Color', srgb: Vector) -> MutableVector:
-        """From sRGB."""
-
-        return cls._from_hsv(parent, HSV._from_srgb(parent, srgb))
-
-    @classmethod
-    def _to_hsl(cls, parent: 'Color', hwb: Vector) -> MutableVector:
-        """To HSL."""
-
-        return HSV._to_hsl(parent, hwb_to_hsv(hwb))
-
-    @classmethod
-    def _from_hsl(cls, parent: 'Color', hsl: Vector) -> MutableVector:
-        """From HSL."""
-
-        return hsv_to_hwb(HSV._from_hsl(parent, hsl))
-
-    @classmethod
-    def _to_hsv(cls, parent: 'Color', hwb: Vector) -> MutableVector:
+    def to_base(cls, hwb: Vector) -> MutableVector:
         """To HSV."""
 
         return hwb_to_hsv(hwb)
 
     @classmethod
-    def _from_hsv(cls, parent: 'Color', hsv: Vector) -> MutableVector:
+    def from_base(cls, hsv: Vector) -> MutableVector:
         """From HSV."""
 
         return hsv_to_hwb(hsv)