+++ coloraide/spaces/jzczhz.py
@@ -4,15 +4,11 @@ JzCzhz class.
 https://www.osapublishing.org/oe/fulltext.cfm?uri=oe-25-13-15131&id=368272
 """
 from ..spaces import Space, RE_DEFAULT_MATCH, GamutUnbound, Lchish, FLG_ANGLE, FLG_OPT_PERCENT
-from .jzazbz import Jzazbz
 from .. import util
 import re
 import math
 from ..util import Vector, MutableVector
-from typing import Tuple, TYPE_CHECKING
-
-if TYPE_CHECKING:  # pragma: no cover
-    from ..color import Color
+from typing import Tuple
 
 ACHROMATIC_THRESHOLD = 0.0003
 
@@ -58,6 +54,7 @@ class JzCzhz(Lchish, Space):
     https://www.osapublishing.org/oe/fulltext.cfm?uri=oe-25-13-15131&id=368272
     """
 
+    BASE = "jzazbz"
     SPACE = "jzczhz"
     SERIALIZE = ("--jzczhz",)
     CHANNEL_NAMES = ("jz", "cz", "hz", "alpha")
@@ -127,25 +124,13 @@ class JzCzhz(Lchish, Space):
         return "hz"
 
     @classmethod
-    def _to_jzazbz(cls, parent: 'Color', jzczhz: Vector) -> MutableVector:
+    def to_base(cls, jzczhz: Vector) -> MutableVector:
         """To Jzazbz."""
 
         return jzczhz_to_jzazbz(jzczhz)
 
     @classmethod
-    def _from_jzazbz(cls, parent: 'Color', jzazbz: Vector) -> MutableVector:
+    def from_base(cls, jzazbz: Vector) -> MutableVector:
         """From Jzazbz."""
 
         return jzazbz_to_jzczhz(jzazbz)
-
-    @classmethod
-    def _to_xyz(cls, parent: 'Color', jzczhz: Vector) -> MutableVector:
-        """To XYZ."""
-
-        return Jzazbz._to_xyz(parent, cls._to_jzazbz(parent, jzczhz))
-
-    @classmethod
-    def _from_xyz(cls, parent: 'Color', xyz: Vector) -> MutableVector:
-        """From XYZ."""
-
-        return cls._from_jzazbz(parent, Jzazbz._from_xyz(parent, xyz))