+++ coloraide/spaces/luv.py
@@ -5,14 +5,9 @@ https://en.wikipedia.org/wiki/CIELUV
 """
 from ..spaces import Space, RE_DEFAULT_MATCH, GamutUnbound, FLG_PERCENT, WHITES, Labish
 from .lab.base import KAPPA, EPSILON, KE
-from .xyz import XYZ
 from .. import util
 import re
 from ..util import Vector, MutableVector
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:  # pragma: no cover
-    from ..color import Color
 
 
 def xyz_to_luv(xyz: Vector, white: str) -> MutableVector:
@@ -111,13 +106,13 @@ class Luv(Labish, Space):
         self._coords[2] = self._handle_input(value)
 
     @classmethod
-    def _to_xyz(cls, parent: 'Color', luv: Vector) -> MutableVector:
+    def to_base(cls, luv: Vector) -> MutableVector:
         """To XYZ."""
 
-        return parent.chromatic_adaptation(cls.WHITE, XYZ.WHITE, luv_to_xyz(luv, cls.WHITE))
+        return luv_to_xyz(luv, cls.WHITE)
 
     @classmethod
-    def _from_xyz(cls, parent: 'Color', xyz: Vector) -> MutableVector:
+    def from_base(cls, xyz: Vector) -> MutableVector:
         """From XYZ."""
 
-        return xyz_to_luv(parent.chromatic_adaptation(XYZ.WHITE, cls.WHITE, xyz), cls.WHITE)
+        return xyz_to_luv(xyz, cls.WHITE)