+++ coloraide/spaces/okhsl.py
@@ -26,16 +26,13 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 from ..spaces import Space, RE_DEFAULT_MATCH, FLG_ANGLE, FLG_OPT_PERCENT, GamutBound, Cylindrical
-from .oklab.base import Oklab, oklab_to_linear_srgb
+from .oklab.base import oklab_to_linear_srgb
 from .. import util
 import re
 import math
 import sys
 from ..util import Vector, MutableVector
-from typing import Tuple, Optional, TYPE_CHECKING
-
-if TYPE_CHECKING:  # pragma: no cover
-    from ..color import Color
+from typing import Tuple, Optional
 
 FLT_MAX = sys.float_info.max
 
@@ -419,6 +416,7 @@ def oklab_to_okhsl(lab: Vector) -> MutableVector:
 class Okhsl(Cylindrical, Space):
     """HSL class."""
 
+    BASE = "oklab"
     SPACE = "okhsl"
     SERIALIZE = ("--okhsl",)
     CHANNEL_NAMES = ("h", "s", "l", "alpha")
@@ -482,49 +480,13 @@ class Okhsl(Cylindrical, Space):
         return coords, alpha
 
     @classmethod
-    def _to_srgb(cls, parent: 'Color', okhsl: Vector) -> MutableVector:
-        """To sRGB."""
-
-        return Oklab._to_srgb(parent, cls._to_oklab(parent, okhsl))
-
-    @classmethod
-    def _from_srgb(cls, parent: 'Color', srgb: Vector) -> MutableVector:
-        """From sRGB."""
-
-        return cls._from_oklab(parent, Oklab._from_srgb(parent, srgb))
-
-    @classmethod
-    def _to_srgb_linear(cls, parent: 'Color', okhsl: Vector) -> MutableVector:
-        """To sRGB Linear."""
-
-        return Oklab._to_srgb_linear(parent, cls._to_oklab(parent, okhsl))
-
-    @classmethod
-    def _from_srgb_linear(cls, parent: 'Color', srgbl: Vector) -> MutableVector:
-        """From SRGB Linear."""
-
-        return cls._from_oklab(parent, Oklab._from_srgb_linear(parent, srgbl))
-
-    @classmethod
-    def _to_oklab(cls, parent: 'Color', okhsl: Vector) -> MutableVector:
+    def to_base(cls, okhsl: Vector) -> MutableVector:
         """To Oklab."""
 
         return okhsl_to_oklab(okhsl)
 
     @classmethod
-    def _from_oklab(cls, parent: 'Color', oklab: Vector) -> MutableVector:
+    def from_base(cls, oklab: Vector) -> MutableVector:
         """From Oklab."""
 
         return oklab_to_okhsl(oklab)
-
-    @classmethod
-    def _to_xyz(cls, parent: 'Color', okhsl: Vector) -> MutableVector:
-        """To XYZ."""
-
-        return Oklab._to_xyz(parent, cls._to_oklab(parent, okhsl))
-
-    @classmethod
-    def _from_xyz(cls, parent: 'Color', xyz: Vector) -> MutableVector:
-        """From XYZ."""
-
-        return cls._from_oklab(parent, Oklab._from_xyz(parent, xyz))