+++ coloraide/spaces/srgb/base.py
@@ -1,42 +1,9 @@
 """SRGB color class."""
 from ...spaces import RE_DEFAULT_MATCH, Space, GamutBound, FLG_OPT_PERCENT
-from ..xyz import XYZ
 from ... import util
 from ...util import Vector, MutableVector
 import re
 import math
-from typing import cast, TYPE_CHECKING
-
-if TYPE_CHECKING:  # pragma: no cover
-    from ...color import Color
-
-RGB_TO_XYZ = [
-    [0.4123907992659593, 0.357584339383878, 0.18048078840183432],
-    [0.21263900587151024, 0.715168678767756, 0.07219231536073373],
-    [0.01933081871559182, 0.11919477979462598, 0.9505321522496608]
-]
-
-XYZ_TO_RGB = [
-    [3.2409699419045226, -1.537383177570094, -0.49861076029300355],
-    [-0.9692436362808796, 1.8759675015077202, 0.04155505740717562],
-    [0.055630079696993635, -0.2039769588889765, 1.0569715142428784]
-]
-
-
-def lin_srgb_to_xyz(rgb: Vector) -> MutableVector:
-    """
-    Convert an array of linear-light sRGB values to CIE XYZ using sRGB's own white.
-
-    D65 (no chromatic adaptation)
-    """
-
-    return cast(MutableVector, util.dot(RGB_TO_XYZ, rgb))
-
-
-def xyz_to_lin_srgb(xyz: Vector) -> MutableVector:
-    """Convert XYZ to linear-light sRGB."""
-
-    return cast(MutableVector, util.dot(XYZ_TO_RGB, xyz))
 
 
 def lin_srgb(rgb: Vector) -> MutableVector:
@@ -50,10 +17,10 @@ def lin_srgb(rgb: Vector) -> MutableVector:
     for i in rgb:
         # Mirror linear nature of algorithm on the negative axis
         abs_i = abs(i)
-        if abs_i < 0.04045:
-            result.append(i / 12.92)
-        else:
+        if abs_i > 0.04045:
             result.append(math.copysign(((abs_i + 0.055) / 1.055) ** 2.4, i))
+        else:
+            result.append(i / 12.92)
     return result
 
 
@@ -78,6 +45,7 @@ def gam_srgb(rgb: Vector) -> MutableVector:
 class SRGB(Space):
     """SRGB class."""
 
+    BASE = "srgb-linear"
     SPACE = "srgb"
     # In addition to the current gamut, check HSL as it is much more sensitive to small
     # gamut changes. This is mainly for a better user experience. Colors will still be
@@ -134,13 +102,13 @@ class SRGB(Space):
         self._coords[2] = self._handle_input(value)
 
     @classmethod
-    def _to_xyz(cls, parent: 'Color', rgb: Vector) -> MutableVector:
-        """SRGB to XYZ."""
+    def from_base(cls, rgb: Vector) -> MutableVector:
+        """Linear sRGB to sRGB."""
 
-        return parent.chromatic_adaptation(cls.WHITE, XYZ.WHITE, lin_srgb_to_xyz(lin_srgb(rgb)))
+        return gam_srgb(rgb)
 
     @classmethod
-    def _from_xyz(cls, parent: 'Color', xyz: Vector) -> MutableVector:
-        """XYZ to SRGB."""
+    def to_base(cls, rgb: Vector) -> MutableVector:
+        """sRGB to linear sRGB."""
 
-        return gam_srgb(xyz_to_lin_srgb(parent.chromatic_adaptation(XYZ.WHITE, cls.WHITE, xyz)))
+        return lin_srgb(rgb)